+++ kibot/out_pcbdraw.py
@@ -303,3 +303,9 @@ class PcbDraw(BaseOutput):  # noqa: F821
         with document:
             self.options = PcbDrawOptions
             """ [dict] Options for the `pcbdraw` output """
+
+    def get_dependencies(self):
+        files = super().get_dependencies()
+        if isinstance(self.options.style, str):
+            files.append(self.options.style)
+        return files